+++ src/formats/table-main-format.ts
@@ -1,6 +1,6 @@
 import type TypeScroll from 'quill/blots/scroll';
 import type { TableValue } from '../utils';
-import { blotName, tableUpSize } from '../utils';
+import { blotName, randomId, tableUpSize } from '../utils';
 import { ContainerFormat } from './container-format';
 import { TableCellInnerFormat } from './table-cell-inner-format';
 import { TableColFormat } from './table-col-format';
@@ -291,6 +291,68 @@ export class TableMainFormat extends ContainerFormat {
       // eslint-disable-next-line unicorn/no-array-for-each
       this.children.forEach(child => child.remove());
       tableMain.moveChildren(this);
+    }
+  }
+
+  // insert row at index
+  insertRow(targetIndex: number) {
+    // get all column id. exclude the columns of the target index row with rowspan
+    const colIds = this.getColIds();
+    const rows = this.descendants(TableRowFormat);
+    const insertColIds = new Set(colIds);
+    let index = 0;
+    for (const row of rows) {
+      if (index === targetIndex) break;
+      row.foreachCellInner((cell) => {
+        if (index + cell.rowspan > targetIndex) {
+          cell.rowspan += 1;
+          insertColIds.delete(cell.colId);
+          // colspan cell need remove all includes colId
+          if (cell.colspan !== 1) {
+            const colIndex = colIds.indexOf(cell.colId);
+            for (let i = 0; i < cell.colspan - 1; i++) {
+              insertColIds.delete(colIds[colIndex + i + 1]);
+            }
+          }
+        }
+      });
+      index += 1;
+    }
+    // append new row
+    const tableId = this.tableId;
+    const rowId = randomId();
+    const tableRow = this.scroll.create(blotName.tableRow, {
+      tableId,
+      rowId,
+    }) as ContainerFormat;
+    for (const colId of insertColIds) {
+      const breakBlot = this.scroll.create('break');
+      const block = breakBlot.wrap('block');
+      const tableCellInner = block.wrap(blotName.tableCellInner, {
+        tableId,
+        rowId,
+        colId,
+        rowspan: 1,
+        colspan: 1,
+      });
+      const tableCell = tableCellInner.wrap(blotName.tableCell, {
+        tableId,
+        rowId,
+        colId,
+        rowspan: 1,
+        colspan: 1,
+      });
+      tableRow.appendChild(tableCell);
+    }
+
+    // insert the new row at the target index
+    // if you insert a row at the thead, then the new row will be inserted in thead. Similarly for tbody and tfoot
+    const refRow = rows[targetIndex] || null;
+    if (!refRow) {
+      rows[rows.length - 1].parent.appendChild(tableRow);
+    }
+    else {
+      refRow.parent.insertBefore(tableRow, refRow);
     }
   }
 }
